🧪 [TEST] Untested function: escapeRegExp
