+++ src/tools/helpers/strings.ts
@@ -33,3 +33,10 @@ export function parseCommaSeparatedList(str: string): string[] {
 
   return result
 }
+
+/**
+ * Escape special characters in a string for use in a regular expression
+ */
+export function escapeRegExp(string: string): string {
+  return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
+}
