+++ tests/helpers/strings.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { parseCommaSeparatedList } from '../../src/tools/helpers/strings.js'
+import { escapeRegExp, parseCommaSeparatedList } from '../../src/tools/helpers/strings.js'
 
 describe('strings helpers', () => {
   describe('parseCommaSeparatedList', () => {
@@ -33,6 +33,33 @@ describe('strings helpers', () => {
 
     it('should handle items with inner spaces', () => {
       expect(parseCommaSeparatedList('word1 word2, word3 word4')).toEqual(['word1 word2', 'word3 word4'])
+    })
+
+    it('should return an empty array for undefined or null-like values (via type safety check)', () => {
+      // @ts-ignore
+      expect(parseCommaSeparatedList(null)).toEqual([])
+      // @ts-ignore
+      expect(parseCommaSeparatedList(undefined)).toEqual([])
+    })
+  })
+
+  describe('escapeRegExp', () => {
+    it('should handle empty strings', () => {
+      expect(escapeRegExp('')).toBe('')
+    })
+
+    it('should return plain strings as-is', () => {
+      expect(escapeRegExp('hello123')).toBe('hello123')
+    })
+
+    it('should escape all regex special characters', () => {
+      const specialChars = '.*+?^' + '${' + '}()|[]\\'
+      const expected = '\\.\\*\\+\\?\\^\\$\\{\\}\\(\\)\\|\\[\\]\\\\'
+      expect(escapeRegExp(specialChars)).toBe(expected)
+    })
+
+    it('should escape special characters mixed with plain text', () => {
+      expect(escapeRegExp('node.name[1]')).toBe('node\\.name\\[1\\]')
     })
   })
 })
